hmac_isal_verify_multi: return bitmask instead of 0/nonzero
- Return type changed from int to uint64_t
- Bit i = 1 means macs[i] does NOT match
- Return 0 means all valid; caller can inspect individual bits
- Fix buffer overflow: ptrs[] VLA sized to num_packets instead of
  HMAC_ISAL_MAX_BATCH
- UINT64_MAX returned on allocation failure (all bits set)
- Test: verify exact bitmask (only bit 1 set) on forged-MAC test
- Add missing <openssl/hmac.h> include and OPENSSL_SUPPRESS_DEPRECATED

+++ include/hmac_isal.h
@@ -87,16 +87,30 @@ int hmac_isal_verify_single(const uint8_t *key, size_t key_len,
                             const hmac_isal_key_cache_t *cache);
 
 /*
- * Constant-time multi-packet HMAC verification.
+ * Constant-time multi-packet HMAC verification (bitmask result).
  *
- * Returns 0 if every macs[i] matches HMAC(key, msgs[i]), nonzero if any
- * differs.  num_packets may be any positive integer; batches larger than
- * 16 are split automatically under the hood.
+ * Returns a bitmask where bit i (LSB = packet 0) is set when
+ * macs[i] does NOT match HMAC(key, msgs[i]).  A return value of 0
+ * means all packets verified successfully.
+ *
+ * The caller can inspect individual bits to decide how to handle
+ * each packet, e.g.:
+ *
+ *   uint64_t bad = hmac_isal_verify_multi(key, klen, msgs, lens, macs, n, NULL);
+ *   if (bad) {
+ *       for (int i = 0; i < n; i++)
+ *           if (bad & ((uint64_t)1 << i))
+ *               handle_bad_packet(i);
+ *   }
+ *
+ * num_packets may be any positive integer; batches larger than 16
+ * are split automatically under the hood.  Only the bits corresponding
+ * to actual packets are significant.
  */
-int hmac_isal_verify_multi(const uint8_t *key, size_t key_len,
-                           const uint8_t *msgs[], const size_t msg_lens[],
-                           const uint8_t *macs[], int num_packets,
-                           const hmac_isal_key_cache_t *cache);
+uint64_t hmac_isal_verify_multi(const uint8_t *key, size_t key_len,
+                                const uint8_t *msgs[], const size_t msg_lens[],
+                                const uint8_t *macs[], int num_packets,
+                                const hmac_isal_key_cache_t *cache);
 
 #ifdef __cplusplus
 }
